new project and solution files
