+++ mssql-schema-dump.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mssql-schema-dump", "mssql-schema-dump\mssql-schema-dump.csproj", "{7C88BB9A-2A66-4A31-8904-74931F8FA867}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x86 = Debug|x86
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7C88BB9A-2A66-4A31-8904-74931F8FA867}.Debug|x86.ActiveCfg = Debug|x86
+		{7C88BB9A-2A66-4A31-8904-74931F8FA867}.Debug|x86.Build.0 = Debug|x86
+		{7C88BB9A-2A66-4A31-8904-74931F8FA867}.Release|x86.ActiveCfg = Release|x86
+		{7C88BB9A-2A66-4A31-8904-74931F8FA867}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+EndGlobal
